third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-computed-rgb.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Color Level 4: Computation of colors using RGB notation</title>
<link rel="help" href="https://drafts.csswg.org/css-color-4/#rgb-functions">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#resolving-sRGB-values">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#serializing-sRGB-values">
<link rel="author" title="Chris Nardi" href="mailto:cnardi@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<style>
  #container {
    color: rgb(255, 0, 0);
  }
</style>
<script>
test_computed_value("color", "rgb(none none none)", "rgb(0, 0, 0)");
test_computed_value("color", "rgb(none none none / none)", "rgba(0, 0, 0, 0)");
test_computed_value("color", "rgb(128 none none)", "rgb(128, 0, 0)");
test_computed_value("color", "rgb(128 none none / none)", "rgba(128, 0, 0, 0)");
test_computed_value("color", "rgb(none none none / .5)", "rgba(0, 0, 0, 0.5)");
test_computed_value("color", "rgb(20% none none)", "rgb(51, 0, 0)");
test_computed_value("color", "rgb(20% none none / none)", "rgba(51, 0, 0, 0)");
test_computed_value("color", "rgb(none none none / 50%)", "rgba(0, 0, 0, 0.5)");
test_computed_value("color", "rgba(none none none)", "rgb(0, 0, 0)");
test_computed_value("color", "rgba(none none none / none)", "rgba(0, 0, 0, 0)");
test_computed_value("color", "rgba(128 none none)", "rgb(128, 0, 0)");
test_computed_value("color", "rgba(128 none none / none)", "rgba(128, 0, 0, 0)");
test_computed_value("color", "rgba(none none none / .5)", "rgba(0, 0, 0, 0.5)");
test_computed_value("color", "rgba(20% none none)", "rgb(51, 0, 0)");
test_computed_value("color", "rgba(20% none none / none)", "rgba(51, 0, 0, 0)");
test_computed_value("color", "rgba(none none none / 50%)", "rgba(0, 0, 0, 0.5)");

test_computed_value("color", "rgb(2.5, 3.4, 4.6)", ["rgb(3, 3, 5)", "rgb(2.5, 3.4, 4.6)"], "[Tests that RGB channels are rounded appropriately]");
</script>
</body>
</html>
